pyproject.toml
[project]
name = "imgdds"
version = "0.1.0"
description = "Convert images to/from DDS format using a simple CLI."
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pillow",
    "imageio[ffmpeg]",
]

[project.scripts]
imgdds = "imgdds.imgdds:main"
